Move relay-group-member-unassign to its own command
